+++ test/config.js
@@ -10,7 +10,8 @@ describe('configuration options', () => {
 			undefined, // eslint-disable-line no-undefined
 			{},
 			{cgi: {fast: 'on'}},
-			{pathAlias: {alias: 'r', procedure: 'p'}}
+			{pathAlias: {alias: 'r', procedure: 'p'}},
+			{errorStyle: 'debug'}
 		].forEach(options => {
 			assert.strictEqual(typeof validate(options), 'object', JSON.stringify(options));
 		});
@@ -28,6 +29,8 @@ describe('configuration errors', () => {
 		{config: {cgi: ''}, error: 'The option "cgi" must be an object where all keys and values are of type string'},
 		{config: {cgi: {a: 1}}, error: 'The option "cgi" must be an object where all keys and values are of type string'},
 		{config: {pathAlias: {alias: 'r', procedure: ''}}, error: 'The option "pathAlias" must be an object with the non-empty string properties "alias" and "procedure"'},
+		{config: {errorStyle: ''}, error: 'The optional option "errorStyle" must be "basic" or "debug"'},
+		{config: {errorStyle: 'on'}, error: 'The optional option "errorStyle" must be "basic" or "debug"'},
 		{config: {trace: true}, error: 'The optional option "trace" must be "on" or "off"'},
 		{config: {trace: ''}, error: 'The optional option "trace" must be "on" or "off"'},
 		{config: {trace: 'enabled'}, error: 'The optional option "trace" must be "on" or "off"'},