+++ CHANGELOG.md
@@ -14,6 +14,19 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 
 
+## [0.1.0] - 2018-09-03
+
+### Added
+- Added the configuration open "errorStyle" that can be set to "basic" or "debug".
+- The current web_plsql version is now a string property "version" in the middleware object.
+- Added the express status monitor to the sample. (http://localhost:8000/status)
+
+### Fixed
+- When downloading a file, first send any eventual "other headers" like "Content-Disposition".
+- Changed x-db-xontent-length to x-db-content-length.
+- Updated all dependencies.
+
+
 ## [0.0.3] - 2018-03-20
 
 ### Added